Move BlockingTestRunnable to the public API space
This is used by both artifacts.  It's simple API should be simple to have used outside of the project without having much expansion to the size.
